requirements.txt
Django==1.8
wheel==0.26.0
djangorestframework==3.3.2